refactor(client): route shared-component imports through the barrel (#56)
Mechanical, generated by a codemod: 45 import lines across 25 files
become 25, all naming `../../shared` instead of the four
`components/<Name>` compatibility shims. Where a file used several
shims the lines collapse into one.

No shim is deleted yet, so this commit stands alone: both paths still
resolve and nothing outside the import lines changes.
