+++ client/src/features/analytics/pages/AdvancedAnalytics.tsx
@@ -3,8 +3,7 @@ import { useTranslation } from '../../../shared/i18n/index';
 import { formatCurrency, formatDate } from '../../../shared/lib/utils';
 import { useApiQuery } from '../../../shared/lib/apiQuery';
 import { Card, CardBody, Button } from '@heroui/react';
-import { Badge } from '../../../shared/components/StatusBadge';
-import PageHeader from '../../../shared/components/PageHeader';
+import { Badge, PageHeader } from '../../../shared';
 import AbcChart from '../components/charts/AbcChart';
 import HeatmapChart from '../components/charts/HeatmapChart';
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, ResponsiveContainer } from 'recharts';
