+++ client/src/features/analytics/pages/Exports.tsx
@@ -3,8 +3,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Download, FileSpreadsheet } from 'lucide-react';
 import toast from 'react-hot-toast';
 import { Button, Select, SelectItem, Card, CardBody } from '@heroui/react';
-import { Badge } from '../../../shared/components/StatusBadge';
-import PageHeader from '../../../shared/components/PageHeader';
+import { Badge, PageHeader } from '../../../shared';
 import { useTranslation } from '../../../shared/i18n/index';
 import { exportToExcel } from '../../../shared/lib/exportUtils';
 import { useApiQuery } from '../../../shared/lib/apiQuery';
