+++ client/src/features/pos/pages/Shifts.tsx
@@ -3,8 +3,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { LogIn, LogOut, Coffee, Play } from 'lucide-react';
 import toast from 'react-hot-toast';
 import { Button, Card, CardBody, Tabs, Tab, Pagination } from '@heroui/react';
-import { Badge } from '../../../shared/components/StatusBadge';
-import PageHeader from '../../../shared/components/PageHeader';
+import { Badge, PageHeader } from '../../../shared';
 import { useTranslation } from '../../../shared/i18n/index';
 import { useAuthStore } from '../../auth';
 import { resource } from '../../../shared/lib/resource';
